Generate steps concurrently.
Before these changes, the step generator was not capable of running
concurrently due to the use of shared state. These changes enable
concurrent step generation by protecting shared state with locks where
necessary.

By far the most subtle changes are those required for the safe
concurrent evaluation of delete-before-replace graphs. Because step
generation is stateful, it is not safe for two processes to concurrently
evaluate steps for the same resource. To prevent this situation, each
resource is locked before its steps are planned and unlocked after its
steps are planned. During the evaluation of a DBR graph, all resources
that _might_ be deleted are locked prior to evaluation and unlocked once
evaluation has completed.

Fixes #3302.